Separate plots from report template
This takes out the code for creating plots from the report template as
it may also be useful in other contexts. In addition to that, the time
limit is removed. I would expect the data to either be prefiltered for
the desired timespan or the initial fetch to have an appropriate `since`
parameter. Also, some of the message type filtering was changed because
I encountered some bugs in the process.